Move ethereum and util tests to /src
