+++ packages/noir-ethereum-api/src/util/object.test.ts
@@ -1,4 +1,4 @@
-import { updateNestedField } from '../../src/util/object.js';
+import { updateNestedField } from './object.js';
 import { describe, it, expect } from 'vitest';
 
 describe('updateNestedField', () => {
